add support for granting temporary partial developer only command access, make g-mute and g-unmute okay by only allowing them only in specialGuilds, rename giveadmin to givedeveloper and minor bugfixes on it, make offshoot command called giveaddlperm for addlbotperms

+++ common/handlers/events/message.js
@@ -35,7 +35,7 @@ module.exports = async msg => {
 
   if (msg.author.bot) return;
 
-  if (mutelist.includes(msg.author.id))
+  if (msg.guild && specialGuilds.includes(msg.guild.id) && mutelist.includes(msg.author.id))
     msg.delete();
 
   if (!msg.guild) {
